tests/requirements_test.txt
# Testing requirements for automated tests
# Install with: pip install -r tests/requirements_test.txt

# Core testing framework
pytest==7.4.3
pytest-django==4.7.0

# Test utilities
pytest-mock==3.12.0
pytest-xdist==3.5.0  # For parallel test execution
pytest-cov==4.1.0    # For coverage reports

# Django testing utilities
factory-boy==3.3.0   # For test data factories
faker==20.1.0        # For generating fake test data

# Assertion helpers (already included with Django but good to be explicit)
# django-extensions  # For additional test utilities if needed

# Development and debugging
pytest-html==4.1.1   # HTML test reports
pytest-json-report==1.5.0  # JSON test reports
